-added whiteboarditemservice -renamed css class postIt to note -persistent whiteboarditem move

+++ de.bht.swp.lao.ocp/src/main/java/de/bht/swp/lao/ocp/attachment/AttachmentService.java
@@ -60,8 +60,8 @@ public class AttachmentService {
 			attachment.setWhiteboard(whiteboardDao.findById(whiteboardid));
 		}
 		
-		attachment.setX(x.intValue());
-		attachment.setY(y.intValue());
+		attachment.setX(x);
+		attachment.setY(y);
 		attachmentDao.save(attachment);
 		
 		Map<String,Object> output = new HashMap<String,Object>();
